move files in pkg/proxy to pkg/proxy/server
