+++ pkg/proxy/server/column.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package proxy
+package server
 
 import (
 	"github.com/pingcap/parser/mysql"
