+++ pkg/proxy/server/packetio.go
@@ -32,7 +32,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package proxy
+package server
 
 import (
 	"bufio"
