+++ pkg/proxy/server/server_util.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package proxy
+package server
 
 import (
 	"sync"
